refactor(tui): split editor and settings patches from pi-tui shim
